requirements.txt
hydra-core>=1.2
pygithub
pandas
tabulate
